refactor: Rename visibility to webgpu
