+++ public/webgpu/Visibility rendering.md
