+++ public/webgpu/main.js
@@ -1,13 +1,13 @@
 import {Vector2, Vector4} from "../../src/math/index.js";
 import {Instance} from "./Instance.js";
-import {VisibilityRenderer} from "./VisibilityRenderer.js";
+import {Renderer} from "./Renderer.js";
 import {listen} from "./input.js";
 
 import {createCamera, createScene} from "./scenes/multipleGeometries.js";
 
 export default async function() {
 	const canvas = document.createElement("canvas");
-	const renderer = new VisibilityRenderer(canvas);
+	const renderer = new Renderer(canvas);
 	const instance = new Instance({
 		renderer,
 		framesPerSecond: 60,
