+++ public/webgpu/scenes/visibility.js
