+++ public/webgpu/shaders/visibility.frag.wgsl
