+++ public/webgpu/shaders/visibility.vert.wgsl
